// app/src/pages/project/metrics/__generated__/ProjectAnnotationMetricsConfigFragment.graphql.ts
export type AnnotationType = "CATEGORICAL" | "CONTINUOUS" | "FREEFORM";
export type OptimizationDirection = "MAXIMIZE" | "MINIMIZE" | "NONE";
import { FragmentRefs } from "relay-runtime";
export type ProjectAnnotationMetricsConfigFragment$data = {
  readonly annotationConfigs: {
    readonly edges: ReadonlyArray<{
      readonly config: {
        readonly annotationType?: AnnotationType;
        readonly lowerBound?: number | null;
        readonly name?: string;
        readonly optimizationDirection?: OptimizationDirection;
        readonly threshold?: number | null;
        readonly upperBound?: number | null;
        readonly values?: ReadonlyArray<{
          readonly label: string;
          readonly score: number | null;
        }>;
      };
    }>;
  };
  readonly " $fragmentType": "ProjectAnnotationMetricsConfigFragment";
};
export type ProjectAnnotationMetricsConfigFragment$key = {
  readonly " $data"?: ProjectAnnotationMetricsConfigFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"ProjectAnnotationMetricsConfigFragment">;
};